Addition of the Insert Method
Addition of the Insert Method, note this is how we will add values to the tree but is not directly linked to those values as we were given specific input methods to use. We will instead declare these stated methods as additional methods that feed the data to the insert method. This has not yet been done but will be done at some point.