Structure & Naming Update
Adds two classes created in the testing file to open the test structure. Also following from a point made in the last retrospective about the potential to show off the project to non-Spartans I have refactored all mentions of Room 3 to read Group 3 to be a little more professional and easily understandable to outsiders.